+++ lib/mutant/parser.rb
@@ -5,6 +5,13 @@ module Mutant
   class Parser
     include Adamantium, Equalizer.new
 
+    class AST
+      include Anima.new(
+        :node,
+        :comment_associations
+      )
+    end
+
     # Initialize object
     #
     # @return [undefined]
@@ -18,7 +25,18 @@ module Mutant
     #
     # @return [AST::Node]
     def call(path)
-      @cache[path] ||= Unparser.parse(path.read)
+      @cache[path] ||= parse(path.read)
+    end
+
+  private
+
+    def parse(source)
+      node, comments = Unparser.parse_with_comments(source)
+
+      AST.new(
+        node:                 node,
+        comment_associations: ::Parser::Source::Comment.associate_by_identity(node, comments)
+      )
     end
 
   end # Parser