+++ spec/unit/mutant/cli_spec.rb
@@ -496,10 +496,8 @@ RSpec.describe Mutant::CLI do
 
       let(:subject_a) do
         Mutant::Subject::Method::Instance.new(
-          context:    Mutant::Context.new(
-            Object,
-            'subject.rb'
-          ),
+          config:     Mutant::Subject::Config::DEFAULT,
+          context:    Mutant::Context.new(Object, 'subject.rb'),
           node:       s(:def, :send, s(:args), nil),
           visibility: :public
         )